requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
decorator==5.1.1
defusedxml==0.7.1
Django==5.0.4
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.19.1
idna==3.7
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
Pygments==2.17.2
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
pyzmq==26.0.2
referencing==0.35.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.2
yarg==0.1.9

setuptools~=68.2.0
thrift~=0.20.0
timestamp~=0.0.1